+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- support multiline changelog entries
+
 ## [1.0.0] - 2025-02-17
 
 ### Changed